Returns metadata about a given client Corresponds to, for example: https://api.municode.com/Clients/name?stateAbbr=VA&clientName=Alexandria

Returns metadata about a given client Corresponds to, for example: https://api.municode.com/Clients/name?stateAbbr=VA&clientName=Alexandria

Usage

get_client_metadata(state_abbreviation, client_name)

Arguments

state_abbreviation

A two-character state code

client_name

The name of a given client

Value

A dataframe of metadata about the specified client

Examples

get_client_metadata("VA", "Alexandria")
#> # A tibble: 1 × 9
#>   client_id client_name state_id state_name state_abbreviation address    city  
#>       <int> <chr>          <int> <chr>      <chr>              <chr>      <chr> 
#> 1       980 Alexandria        46 Virginia   VA                 PO Box 178 Alexa…
#> # ℹ 2 more variables: zip_code <chr>, website <chr>